Added Peppermint in the list of tested distributions.
Peppermint, is an Ubuntu based distribution, tailored to netbooks, with a certain mix of LXDE, Xfce & Cinnamon composing the main desktop environnement. Oh, and it has an SSB manager too. 
Since it doesn't change anything from a base Ubuntu that cause a case of packages incompatibility, this pull request can be safely refused.