fix: fix available memory output for NetBSD and use whoami for Android
- NetBSD doesn't expose a sysctl entry for available memory
and the kernel module isn't callable from shell script. Use
vmstat(1) with awk(1) to fix the current issue.

- Replace "USER" variable entry with whoami(1). It seems
some shells don't interpret ${$(...)}, and we can replace
this by using whoami(1). It's available in coreutils, and
most other Unix-like userlands.

- Only redirect stdout to /dev/null as command -v doesn't
give a stderr and only a return value is provided by it.
This fixes printing path output of getprop when running
NerdFetch on Android.